% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHWplacement-package.R
\docType{package}
\name{CHWplacement-package}
\alias{CHWplacement}
\alias{CHWplacement-package}
\title{CHWplacement: What the Package Does (One Line, Title Case)}
\description{
This package contains functions to create optimised placement scenarios for Community Health Workers.
}
\author{
\strong{Maintainer}: Clara Champagne \email{clara.champagne@swisstph.ch}

}
\keyword{internal}
