% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateReadCHWplacement.R
\name{CreateCHWplacement}
\alias{CreateCHWplacement}
\title{Creates the integer program associated with the CHW placement problem}
\usage{
CreateCHWplacement(
  population.raster,
  friction.raster,
  shp,
  name,
  buffer,
  radius,
  capacity.name = "MSPP",
  access.raster,
  popurb.raster,
  max.treat.per.CHW.urban = 2500,
  max.treat.per.CHW.rural = 1000,
  max.CHW.per.pixel,
  is.inside = FALSE,
  filepath
)
}
\arguments{
\item{population.raster}{raster map containing population densities}

\item{friction.raster}{raster map containing friction surface}

\item{shp}{shapefile of the area of interest}

\item{name}{name (character) of the MPS file}

\item{buffer}{buffer around current HF}

\item{radius}{dimension of the radius around the point (same dimension as T.GC, e.g minutes)}

\item{capacity.name}{name (character) given to the capacity definition (for folder naming purposes)}

\item{access.raster}{raster map containing the distance to HF layer}

\item{popurb.raster}{raster map containing the urban population layer (urban pixels show the population in the pixel and the rural pixels are NA). Can be created with \link[CHWplacement]{DefineUrban}.}

\item{max.treat.per.CHW.urban}{max number of people per CHW in urban areas}

\item{max.treat.per.CHW.rural}{max number of people per CHW in rural areas}

\item{max.CHW.per.pixel}{maximum number of CHW that can be placed on the same pixel}

\item{is.inside}{whether CHW are placed inside the buffer (if FALSE, they are placed outside)}

\item{filepath}{directory where to store the MPS file}
}
\value{
a MPS file containing the optimisation program
}
\description{
This function combined the different rasters and shapefiles needed to create the optimisation program
}
