% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateReadCHWplacement_SCIP.R
\name{CreateCHWplacement_SCIP}
\alias{CreateCHWplacement_SCIP}
\title{Creation of MPS file and txt.Script}
\usage{
CreateCHWplacement_SCIP(
  population.raster,
  friction.raster,
  access.raster,
  popurb.raster,
  shp,
  filepath,
  name,
  buffer,
  radius,
  capacity.name = "MSPP",
  max.treat.per.CHW.urban = 2500,
  max.treat.per.CHW.rural = 1000,
  max.CHW.per.pixel,
  is.inside = FALSE,
  dualitygap = 1,
  relativegap = F,
  SCIP_version = "7.0.3"
)
}
\arguments{
\item{population.raster}{raster map containing population densities}

\item{friction.raster}{raster map containing friction surface}

\item{access.raster}{raster map containing the distance to HF layer}

\item{popurb.raster}{raster map containing the urban population layer (urban pixels show the population in the pixel and the rural pixels are NA). Can be created with \link[CHWplacement]{DefineUrban}.}

\item{shp}{shapefile of the area of interest}

\item{filepath}{directory where to store the MPS file}

\item{name}{(character) name of the MPS file}

\item{buffer}{buffer around current HF}

\item{radius}{dimension of the radius around the point (same dimension as T.GC, e.g minutes)}

\item{capacity.name}{name (character) given to the capacity definition (for folder naming purposes)}

\item{max.treat.per.CHW.urban}{max number of people per CHW in urban areas}

\item{max.treat.per.CHW.rural}{max number of people per CHW in rural areas}

\item{max.CHW.per.pixel}{maximum number of CHW that can be placed on the same pixel}

\item{is.inside}{whether CHW are placed inside the buffer (if FALSE, they are placed outside)}

\item{dualitygap}{condition of suboptimality acceptance for the solver:
give an upper bounds of the difference with an hypothetical optimal solution}

\item{relativegap}{if true -> |primal-dual|/MIN(|dual|,|primal|)
else       |primal-dual|    (more interpretable but less accurate)}

\item{SCIP_version}{version of the SCIP shell downloaded}
}
\value{
a MPS file containing the optimisation program and txt script to run it in SCIP shell
}
\description{
This function executes the CLSCP problem to find optimal locations (Current & Storbeck)
using the lpsolveAPI package for integer programming. Relies on the functions CHW_CreatProgScript_SCIP and  PrepareRasterFiles (see functions_CLSCP)
}
