% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateReadCHWplacement_SCIP.R
\name{CreateScriptSCIP}
\alias{CreateScriptSCIP}
\title{Creation of a Script in order to use SCIP on an MPS file}
\usage{
CreateScriptSCIP(
  mps_directory,
  name,
  dualitygap = 0,
  relativegap = T,
  SCIP_version = "7.0.3"
)
}
\arguments{
\item{mps_directory}{directory of the mps file}

\item{name}{name of the mps file}

\item{dualitygap}{condition of suboptimality acceptance for the solver:
give an upper bounds of the difference with an hypothetical optimal solution}

\item{relativegap}{if true -> |primal-dual|/MIN(|dual|,|primal|)
else       |primal-dual|    (more interpretable but less accurate)}

\item{SCIP_version}{usefull for creating SCIP_path so only for windows user}
}
\description{
This function use Paths informations to produce a txt file which containes two Script
A manual one which can be use in every OS and version, and an Automated one in Ps1 (Windows script)
If the Windows script don't work, verify the Script version and maybe change the "SCIP_path" in the function.
}
