% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineUrbanRaster.R
\name{define_urban}
\alias{define_urban}
\title{Defines rural and urban areas based on population density, using the terra package}
\usage{
define_urban(
  population_raster,
  rururb_cutoff = 300,
  min_urbsize = 2000,
  directions = 8,
  mask = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{population_raster}{An input raster containing people per pixel. Default inputs assume input resolution to be approximately 1 sq. km. resolution.}

\item{rururb_cutoff}{Minimum population per pixel to be eligible for urban classification.}

\item{min_urbsize}{Minimum total population of clustered pixels to the classified as urban.}

\item{directions}{See terra::patches, default is 8.}

\item{mask}{TRUE/FALSE, Should the returned raster have masked values (1 = Urban, 0 = Rural). Default is FALSE.}

\item{verbose}{TRUE/FALSE Print messages?}
}
\value{
A SpatRaster object similar to population.raster in which all non-urban pixels are masked (i.e. NA), or a categorized raster where 0 = Rural and 1 = Urban if mask = TRUE.
}
\description{
This function defines urban areas as those with a population density
higher than rururb_cutoff per pixel. Urban areas are also contiguous and have a minimal
total population equal to min_urbsize. The methodology is derived from \url{http://documents.worldbank.org/curated/en/709121516634280180/pdf/122880-V1-WP-P156561-OUO-9-FINAL-ENGLISH.pdf}.
This function differs from \emph{DefineUrban} in that the \emph{terra} package is used (instead of \emph{raster}).
The results between the 2 functions may differ due to different handling of diagonal values in the \emph{terra} and \emph{raster} packages.
}
