% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineUrbanRaster.R
\name{DefineUrban}
\alias{DefineUrban}
\title{Defines rural and urban areas based on population density}
\usage{
DefineUrban(population.raster, rururb_cutoff = 300, min_urbsize = 2000)
}
\arguments{
\item{population.raster}{raster map containing population densities}

\item{rururb_cutoff}{minimal population density to define an urban area}

\item{min_urbsize}{minimal total population size to define an urban area}
}
\value{
a raster similar to population.raster in which all non-urban pixels are masked (i.e. NA)
}
\description{
This function defines urban areas as those with a population density
higher than rururb_cutoff per pixel. Urban areas are also contiguous and have a minimal
total population equal to min_urbsize. The methodology is derived from \url{http://documents.worldbank.org/curated/en/709121516634280180/pdf/122880-V1-WP-P156561-OUO-9-FINAL-ENGLISH.pdf}
}
