% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_footprint_pixel.R
\name{GetFootprintWalkingTime}
\alias{GetFootprintWalkingTime}
\title{Gets the footprint around a pixel}
\usage{
GetFootprintWalkingTime(pos, radius, pop.map, T.GC)
}
\arguments{
\item{pos}{coordinates of the given points}

\item{radius}{dimension of the radius around the point (same dimension as T.GC, e.g minutes)}

\item{pop.map}{raster map containing population densities}

\item{T.GC}{transition matrix T.CG associated to the map pop.map}
}
\value{
a raster
}
\description{
This function returns the coordinates of all the points in the surrounding of a
given list of spatial points. The surrounding is based on walking distance within
radius, using the transition matrix T.CG associated to the map pop.map
}
