% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_CLSCP.R
\name{OptiCLSCPWalkingTimeMPS}
\alias{OptiCLSCPWalkingTimeMPS}
\title{Creates the integer program associated with the CLSCP problem}
\usage{
OptiCLSCPWalkingTimeMPS(
  pop.map,
  access.raster,
  radius,
  T.GC,
  popurb = "none",
  max.treat.per.CHW.urban,
  max.treat.per.CHW.rural,
  max.CHW.per.pixel = 1,
  directory,
  name
)
}
\arguments{
\item{pop.map}{raster map containing population densities}

\item{access.raster}{raster containing access to HF variable}

\item{radius}{dimension of the radius around the point (same dimension as T.GC, e.g minutes)}

\item{T.GC}{transition matrix T.CG associated to the map pop.map}

\item{popurb}{raster map containing the urban population layer}

\item{max.treat.per.CHW.urban}{max number of people per CHW in urban areas}

\item{max.treat.per.CHW.rural}{max number of people per CHW in rural areas}

\item{max.CHW.per.pixel}{maximum number of CHW that can be placed on the same pixel}

\item{directory}{directory where to store the MPS file}

\item{name}{name of the MPS file}
}
\value{
a MPS file containing the optimisation program
}
\description{
This function executes the CLSCP problem to find optimal locations (Current & Storbeck)
using the lpsolveAPI package for integer programming.
}
