% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-sol.R
\name{parse_gurobi_sol}
\alias{parse_gurobi_sol}
\title{Read Gurobi output}
\usage{
parse_gurobi_sol(
  population.raster,
  friction.raster,
  shp,
  name,
  buffer,
  radius,
  capacity.name,
  access.raster,
  popurb.raster,
  is.inside,
  filepath,
  write = F,
  check = TRUE
)
}
\arguments{
\item{population.raster}{raster map containing population densities}

\item{friction.raster}{raster map containing friction surface}

\item{shp}{shapefile of the area of interest}

\item{name}{name (character) of the .sol file}

\item{buffer}{buffer around current HF}

\item{radius}{dimension of the radius around the point (same dimension as T.GC, e.g minutes)}

\item{capacity.name}{name (character) given to the capacity definition (for folder naming purposes)}

\item{access.raster}{raster map containing the distance to HF layer}

\item{popurb.raster}{raster map containing the urban population layer (urban pixels show the population in the pixel and the rural pixels are NA). Can be created with \link[CHWplacement]{DefineUrban}.}

\item{is.inside}{whether CHW are placed inside the buffer (if FALSE, they are placed outside)}

\item{filepath}{directory where to store the MPS file}

\item{write}{if TRUE, the output is written as a csv file in filepath}

\item{check}{Run sanity checks?}
}
\value{
a data.frame containing details on CHW optimization results
}
\description{
This function tranforms the output of a Gurobi opitimisation into map coordinates
}
