% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateReadCHWplacement.R
\name{PrepareRasterFiles}
\alias{PrepareRasterFiles}
\title{Prepare raster files for CHW placement algorithm}
\usage{
PrepareRasterFiles(
  population.raster,
  friction.raster,
  shp,
  buffer,
  access.raster,
  popurb.raster,
  is.inside
)
}
\arguments{
\item{population.raster}{raster map containing population densities}

\item{friction.raster}{raster map containing friction surface}

\item{shp}{shapefile of the area of interest}

\item{buffer}{buffer around current HF}

\item{access.raster}{raster map containing the distance to HF layer}

\item{popurb.raster}{raster map containing the urban population layer (urban pixels show the population in the pixel and the rural pixels are NA). Can be created with \link[CHWplacement]{DefineUrban}.}

\item{is.inside}{whether CHW are placed inside the buffer (if FALSE, they are placed outside)}
}
\value{
a list containing the pre-processed rasters and transition matrix
}
\description{
This function combined the different rasters and shapefiles needed to create the optimisation program
}
