% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateReadCHWplacement_SCIP.R
\name{read_CHWplacement_sol_SCIP}
\alias{read_CHWplacement_sol_SCIP}
\title{Read SCIP results}
\usage{
read_CHWplacement_sol_SCIP(
  population.raster,
  friction.raster,
  access.raster,
  popurb.raster,
  shp,
  name,
  filepath,
  radius,
  capacity.name,
  buffer,
  is.inside,
  write = F
)
}
\arguments{
\item{population.raster}{raster map containing population densities}

\item{friction.raster}{raster map containing friction surface}

\item{access.raster}{raster map containing the distance to HF layer}

\item{popurb.raster}{raster map containing the urban population layer (urban pixels show the population in the pixel and the rural pixels are NA). Can be created with \link[CHWplacement]{DefineUrban}.}

\item{shp}{shapefile of the area of interest}

\item{name}{name (character) of the MPS file}

\item{filepath}{directory where to store the MPS file}

\item{radius}{dimension of the radius around the point (same dimension as T.GC, e.g minutes)}

\item{capacity.name}{name (character) given to the capacity definition (for folder naming purposes)}

\item{buffer}{buffer around current HF}

\item{is.inside}{whether CHW are placed inside the buffer (if FALSE, they are placed outside)}

\item{write}{if TRUE, the output is written as a csv file in filepath}
}
\value{
a csv file containing the optimisation program
}
\description{
This function tranforms the output of a SCIP opitimisation into map coordinates
}
