% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_R.R
\name{solve_CHWplacement_wrapper}
\alias{solve_CHWplacement_wrapper}
\title{Solve optimisation problem with R (lpSolve API)}
\usage{
solve_CHWplacement_wrapper(
  population.raster,
  friction.raster,
  shp,
  buffer,
  access.raster,
  radius,
  capacity.name,
  popurb.raster,
  name,
  is.inside = F,
  write = F,
  filepath,
  mytimeout = 10
)
}
\arguments{
\item{population.raster}{raster map containing population densities}

\item{friction.raster}{raster map containing friction surface}

\item{shp}{shapefile of the area of interest}

\item{buffer}{buffer around current HF}

\item{access.raster}{raster map containing the distance to HF layer}

\item{radius}{dimension of the radius around the point (same dimension as T.GC, e.g minutes)}

\item{capacity.name}{??}

\item{popurb.raster}{raster map containing the urban population layer}

\item{name}{name of the MPS file}

\item{is.inside}{whether CHW are placed inside the buffer (if FALSE, they are placed outside)}

\item{write}{f TRUE, the output is written as a csv file in filepath}

\item{filepath}{directory where to store the MPS file}

\item{mytimeout}{maximal time (in seconds) that the function is allowed to search for the maximum}
}
\value{
a MPS file containing the optimisation program
}
\description{
This function tranforms the output of a Gurobi opitimisation into map coordinates
}
